Refactored Length to Size in Property. Reworked ReferenceArrayHelper to be more low-level. Updated StringProperty and ReferenceProperties. Updated Glossary Items to support uint (from a single byte) Reworked javascript decimalToHexdecimal and hexdecimalToDecimal functions.

+++ src/GameHook.Domain/BuildEnvironment.cs
@@ -1,5 +1,5 @@
-global using GameHookGlossaryPage = System.Collections.Generic.IDictionary<byte, dynamic>;
-global using GameHookGlossary = System.Collections.Generic.IDictionary<string, System.Collections.Generic.IDictionary<byte, dynamic>>;
+global using GameHookGlossaryPage = System.Collections.Generic.IDictionary<uint, dynamic>;
+global using GameHookGlossary = System.Collections.Generic.IDictionary<string, System.Collections.Generic.IDictionary<uint, dynamic>>;
 global using GameHookMacros = System.Collections.Generic.IDictionary<string, System.Collections.Generic.IDictionary<object, object>>;
 global using MemoryAddress = System.UInt32;
 using System.Reflection;
